# tx-backend/openapi/traceability-foss-backend.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Trace-FOSS - OpenAPI Documentation",
    "description": "Trace-FOSS is a system for tracking parts along the supply chain. A high level of transparency across the supplier network enables faster intervention based on a recorded event in the supply chain. This saves costs by seamlessly tracking parts and creates trust through clearly defined and secure data access by the companies and persons involved in the process.",
    "license": {
      "name": "License: Apache 2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9999/api",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/investigations": {
      "post": {
        "tags": [
          "Investigations"
        ],
        "summary": "Start investigations by part ids",
        "description": "The endpoint starts investigations based on part ids provided.",
        "operationId": "investigateAssets",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartInvestigationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Authorization failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartInvestigationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartInvestigationResponse"
                }
              }
            }
          },
          "201": {
            "description": "Created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartInvestigationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oAuth2": [
              "profile email"
            ]
          }
        ]
      }
    },
    "/investigations/{investigationId}/update": {
      "post": {
        "tags": [
          "Investigations"
        ],
        "summary": "Update investigations by id",
        "description": "The endpoint updates investigations by their id.",
        "operationId": "updateInvestigation",
        "parameters": [
          {
            "name": "investigationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvestigationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Authorization failed."
          },
          "403": {
            "description": "Forbidden."
          },
          "204": {
            "description": "No content."
          }
        },
        "security": [
          {
            "oAuth2": [
              "profile email"
            ]
          }
        ]
      }
    },
    "/investigations/{investigationId}/close": {
      "post": {
        "tags": [
          "Investigations"
        ],
        "summary": "Close investigations by id",
        "description": "The endpoint closes investigations by their id.",
        "operationId": "closeInvestigation",
        "parameters": [
          {
            "name": "investigationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseInvestigationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Authorization failed."
          },
          "403": {
            "description": "Forbidden."
          },
          "204": {
            "description": "No content."
          }
        },
        "security": [
          {
            "oAuth2": [
              "profile email"
            ]
          }
        ]
      }
    },
    "/investigations/{investigationId}/cancel": {
      "post": {
        "tags": [
          "Investigations"
        ],
        "summary": "Cancles investigations by id",
        "description": "The endpoint cancles investigations by their id.",
        "operationId": "cancelInvestigation",
        "parameters": [
          {
            "name": "investigationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Authorization failed."
          },
          "403": {
            "description": "Forbidden."
          },
          "204": {
            "description": "No content."
          }
        },
        "security": [
          {
            "oAuth2": [
              "profile email"
            ]
          }
        ]
      }
    },
    "/investigations/{investigationId}/approve": {
      "post": {
        "tags": [
          "Investigations"
        ],
        "summary": "Approves investigations by id",
        "description": "The endpoint approves investigations by their id.",
        "operationId": "approveInvestigation",
        "parameters": [
          {
            "name": "investigationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Authorization failed."
          },
          "403": {
            "description": "Forbidden."
          },
          "204": {
            "description": "No content."
          }
        },
        "security": [
          {
            "oAuth2": [
              "profile email"
            ]
          }
        ]
      }
    },
...
    "/investigations/{investigationId}": {
      "get": {
        "tags": [
          "Investigations"
        ],
        "summary": "Gets investigations by id",
        "description": "The endpoint returns investigations as paged result by their id.",
        "operationId": "getInvestigation",
        "parameters": [
          {
            "name": "investigationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Authorization failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationData"
                }
              }
            }
          },
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationData"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationData"
                }
              }
            }
          }
        },
        "security": [
          {
            "oAuth2": [
              "profile email"
            ]
          }
        ]
      }
    },
    "/investigations/received": {
      "get": {
        "tags": [
          "Investigations"
        ],
        "summary": "Gets received investigations",
        "description": "The endpoint returns received investigations as paged result.",
        "operationId": "getReceivedInvestigations",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Authorization failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultInvestigationData"
                }
              }
            }
          },
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultInvestigationData"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultInvestigationData"
                }
              }
            }
          }
        },
        "security": [
          {
            "oAuth2": [
              "profile email"
            ]
          }
        ]
      }
    },
    "/investigations/created": {
      "get": {
        "tags": [
          "Investigations"
        ],
        "summary": "Gets created investigations",
        "description": "The endpoint returns created investigations as paged result.",
        "operationId": "getCreatedInvestigations",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Authorization failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultInvestigationData"
                }
              }
            }
          },
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultInvestigationData"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultInvestigationData"
                }
              }
            }
          }
        },
        "security": [
          {
            "oAuth2": [
              "profile email"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StartInvestigationRequest": {
        "required": [
          "severity"
        ],
        "type": "object",
        "properties": {
          "partIds": {
            "maxItems": 100,
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "maxLength": 1000,
            "minLength": 15,
            "type": "string"
          },
          "targetDate": {
            "type": "string",
            "format": "date-time"
          },
          "severity": {
            "type": "string"
          }
        }
      },
      "StartInvestigationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "UpdateInvestigationRequest": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "SENT",
              "RECEIVED",
              "ACKNOWLEDGED",
              "ACCEPTED",
              "DECLINED",
              "CANCELED",
              "CLOSED"
            ]
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 15,
            "type": "string"
          }
        }
      },
      "CloseInvestigationRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "maxLength": 1000,
            "minLength": 15,
            "type": "string"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "InvestigationData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdByName": {
            "type": "string"
          },
          "createdDate": {
            "type": "string"
          },
          "assetIds": {
            "maxItems": 1000,
            "minItems": 0,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "channel": {
            "type": "string",
            "enum": [
              "SENDER",
              "RECEIVER"
            ]
          },
          "reason": {
            "$ref": "#/components/schemas/InvestigationReason"
          },
          "sendTo": {
            "type": "string"
          },
          "sendToName": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "targetDate": {
            "type": "string"
          }
        }
      },
      "InvestigationReason": {
        "type": "object",
        "properties": {
          "close": {
            "type": "string"
          },
          "accept": {
            "type": "string"
          },
          "decline": {
            "type": "string"
          }
        }
      },
      "PageResultInvestigationData": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvestigationData"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalItems": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
...
  }
}